% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_funs.R
\name{cross_fun}
\alias{cross_fun}
\title{Calculation of the voter turnout difference between two vote topics ("Vorlagen")}
\usage{
cross_fun(df, vorl1, vorl2, geo_cols = geocols)
}
\arguments{
\item{df}{A data frame containing the municipality ID and voter turnout for various proposals.
Each column represents a specific proposal, and the column names should correspond
to the proposal IDs (e.g., 'eidg1', 'kant2').}

\item{vorl1}{A character vector specifying the proposal ID for the first proposal (e.g., "eidg1", "kant2").
The proposal ID should match the column name in the dataset.}

\item{vorl2}{A character vector specifying the proposal ID for the second proposal (e.g., "eidg1", "kant2").
This should also correspond to a column name in the dataset.}

\item{geocols}{Geo-Column
This should also correspond to a column name in the dataset.}
}
\value{
Tibble with the turnout difference between two vote topics
}
\description{
Calculation of the voter turnout difference between two vote topics ("Vorlagen")
}
\examples{

testdata <- tibble(gemwkid = c(13,49,41,43,44),
eidg1 = c(60.90,61.18,65.27,55.36,57.68),
eidg2 = c(62.16,62.54,66.95,56.65,58.68),
kant1 = c(57.73,60.27,63.31,51.93,54.49))

cross_fun(testdata,"eidg1","eidg2")

 # generate combinations
combinations <-as.data.frame(t(combn(c("eidg1","eidg2","kant1"),2)))

# difference between columns named as the first combination
cross_fun(crosscheckdata,combinations$V1[1],combinations$V2[1])

}
