% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{double_mad}
\alias{double_mad}
\title{Get the left and right Median Absolute Deviations (MAD) from the median for asymmetric distributions}
\usage{
double_mad(x, zero_mad_action = NULL)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{zero_mad_action}{Determines the action in the event of an MAD of zero.
Defaults to NULL. The options are:
\itemize{
\item \strong{NULL}: process runs with no warning
\item \strong{"warn"}: a warning will be displayed
\item \strong{"stop"}: process is stopped
}}
}
\value{
A numeric vector of length 2.
}
\description{
Suited to find outliers in asymetric distributions (in contrast to the standard mad() function which works for symetric distributions only)
The function splits the values along the median and returns separate MADs for the left and the right side of the distribution.
https://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers/
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

double_mad(x)

}
