% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{double_mad_from_median}
\alias{double_mad_from_median}
\title{Calculate the distance of a value from the median of a distribution in relation to its Median Absolute Deviation (MAD)}
\usage{
double_mad_from_median(x, zero_mad_action = NULL)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{zero_mad_action}{Determines the action in the event of an MAD of zero.
Defaults to NULL. The options are:
\itemize{
\item \strong{NULL}: process runs with no warning
\item \strong{"warn"}: a warning will be displayed
\item \strong{"stop"}: process is stopped
}}
}
\value{
A numeric vector of length \code{length(x)}.
}
\description{
This function is suited to find outliers in asymetric distributions (in contrast to the standard mad() function which works for
symetric distributions only). The function splits the values along the median and returns the distance for every value from the
median, relative to the left or right side MAD.
https://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers/
}
\examples{

x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

double_mad_from_median(x)

}
