% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{DoubleMAD}
\alias{DoubleMAD}
\title{Get Median Absolute Deviation from the Median for asymmetric distributions}
\usage{
DoubleMAD(x, zero.mad.action = "warn")
}
\arguments{
\item{x}{numeric vector}

\item{zero.mad.action}{defaults to "warn" : if MAD = 0 a warning will be displayed}
}
\value{
numeric value
}
\description{
Suited to find outliers in asymetric distributions (in contrast to the standard mad() function which works for symetric distributions only)
https://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers/
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

DoubleMAD(x)

}
