% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{DoubleMADsFromMedian}
\alias{DoubleMADsFromMedian}
\title{Calculate the distance of a value from the Median of a distribution in term of Median Absolute Deviations (MAD)}
\usage{
DoubleMADsFromMedian(x, zero.mad.action = "warn")
}
\arguments{
\item{x}{vector of numeric values}

\item{zero.mad.action}{action in the event of a MAD of zero (Options: "stop", "warn", "na" and "warn and na")}
}
\value{
numeric vector
}
\description{
Calculate the distance of a value from the Median of a distribution in term of Median Absolute Deviations (MAD)
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

DoubleMADsFromMedian(x)
}
