% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{is_outlier_single_mad}
\alias{is_outlier_single_mad}
\title{Detect outliers using Z-score with MAD for symmetric distributions}
\usage{
is_outlier_single_mad(x, threshold = 3, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{threshold}{Z-score threshold (defaults to 3).}

\item{na.rm}{Remove NAs, defaults to TRUE.}
}
\value{
A logical vector.
}
\description{
Outlier detection based on Median Absolute Deviation (MAD) for symetric distributions. The function calculates the distance
to the median for every value in the distribution relative to the MAD. It then compares the value to your threshold and
labels the outliers.
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

is_outlier_single_mad(x)

}
