% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{isnt_out_z}
\alias{isnt_out_z}
\title{Z-score}
\usage{
isnt_out_z(x, thres = 3, na.rm = TRUE)
}
\arguments{
\item{x}{variable of interest}

\item{thres}{z-score threshold (defaults to 3, which is a popular choice).}

\item{na.rm}{remove NAs, defaults to TRUE}
}
\value{
logical vector
}
\description{
Z-score, also called a standard score, of an observation is broadly speaking a distance from the population center measured in number of normalization units. The default choice for center is sample mean and for normalization unit is standard deviation.
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

isnt_out_z(x)

}
