% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{predict_single_vote}
\alias{predict_single_vote}
\title{Run prediction for one vote}
\usage{
predict_single_vote(model, testdata)
}
\arguments{
\item{model}{A trained model, generated using \link[=train_prediction_model]{plausi::train_prediction_model()}.}

\item{testdata}{Dataset on which the prediction should be run. The data must contain all columns of the training
data of the model \code{model$trainingData}.}
}
\value{
A data.frame.
}
\description{
This function can be used to predict the outcome of one vote based on a trained model,
generated by using the \link[=train_prediction_model]{plausi::train_prediction_model()}.
To create replicable examples, use the function together with set.seed().
}
\examples{

# Set seed for reproducibility
set.seed(42)

test_model <- train_prediction_model("Eidg1", vote_data, to_exclude_vars = "Kant1")

predict_single_vote(test_model, vote_data)

}
