% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result_data}
\alias{result_data}
\title{Test data containing results for national votes from the canton of Zurich for
all Sundays from 2017-03-01 to 2020-09-27.}
\format{
A dataframe with 3864 rows and 16 columns.
\describe{
\item{name}{Name of the referendum.}
\item{id}{ID of the referendum.}
\item{canton_id}{ID of the canton.}
\item{canton_name}{Name of the canton.}
\item{mun_name}{Name of the municipality.}
\item{geoLevelParentnummer}{Geo level number of the parent geo unit, in the case of municipalities that means district number.}
\item{gebietAusgezaehlt}{Indicator for finalised counting status.}
\item{jaStimmenInProzent}{Percentage of yes-votes.}
\item{jaStimmenAbsolut}{Absolut number of yes-votes.}
\item{neinStimmenAbsolut}{Absolut number of no-votes.}
\item{stimmbeteiligungInProzent}{Turnout.}
\item{eingelegteStimmzettel}{Total number of submitted ballots.}
\item{anzahlStimmberechtigte}{Number of elligable voters.}
\item{gueltigeStimmen}{Total number of valid yes- and no-votes.}
\item{votedate}{Date of the vote.}
}
}
\usage{
result_data
}
\description{
Test data containing results for national votes from the canton of Zurich for
all Sundays from 2017-03-01 to 2020-09-27.
}
\keyword{datasets}
