% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{rmse}
\alias{rmse}
\title{Calculate RMSE}
\usage{
rmse(prediction, observation, na.rm = TRUE)
}
\arguments{
\item{prediction}{Predicted value.}

\item{observation}{Oserved value.}

\item{na.rm}{Remove NA values, defaults to TRUE}
}
\value{
A vector of numeric values.
}
\description{
Calculate the Root Mean Square Error (RMSE). The RMSE is the standard deviation of the residuals (prediction
errors) and therefore an indicator of how precise the prediction of a specific vote actually is.
}
\examples{

pred_data <- predict_votes(c("Eidg1", "Kant1"), vote_data, exclude_votes = TRUE)

pred_data$rmse <- rmse(pred_data$pred, pred_data$real)

}
