% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{rmse_cutoff}
\alias{rmse_cutoff}
\title{Calculate RMSE after excluding the most extreme values}
\usage{
rmse_cutoff(
  prediction,
  observation,
  cutoff_min = 2,
  cutoff_perc = 5,
  na.rm = TRUE
)
}
\arguments{
\item{prediction}{Predicted value.}

\item{observation}{Oserved value.}

\item{cutoff_min}{Numeric. Minimum absolute number of highest residuals to be cut off before calculating the RMSE.}

\item{cutoff_perc}{Numeric. Percentage of highest residuals to be cut off before calculating the RMSE.}

\item{na.rm}{Remove NA values, defaults to TRUE}
}
\value{
A vector of numeric values.
}
\description{
Calculate the Root Mean Square Error (RMSE) after the exclusion of the most extreme values.
The RMSE is the standard deviation of the residuals (prediction errors) and therefore an
indicator of how precise the prediction of a specific vote actually is.
}
\examples{

# Set seed for reproducibility
set.seed(42)

pred_data <- predict_votes(c("Eidg1", "Kant1"), vote_data, exclude_votes = TRUE)

pred_data$rmse <- rmse(pred_data$pred, pred_data$real)

}
