% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{train_prediction_model}
\alias{train_prediction_model}
\title{Train model for prediction of one vote}
\usage{
train_prediction_model(
  x,
  traindata,
  method = "svmRadial",
  trControl = NULL,
  to_exclude_vars = NULL,
  geovars = c("gemeinde", "v_gemwkid"),
  training_prop = NA,
  ...
)
}
\arguments{
\item{x}{Column name of the dependent variable.}

\item{traindata}{Data used to train the model containing the dependent variable and the predictor columns.}

\item{method}{A string specifying which classification or
regression model to use. Possible values are found using
\code{names(getModelInfo())}. See
\url{http://topepo.github.io/caret/train-models-by-tag.html}. A
list of functions can also be passed for a custom model
function. See
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}
for details.}

\item{trControl}{A list of values that define how this function
acts. See \code{\link[caret]{trainControl}} and
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}.
(NOTE: If given, this argument must be named.)}

\item{to_exclude_vars}{Variables that should be excluded from the model. It makes sense to exclude other
votes from the current Sunday since these can contain a lot of NAs that negatively impact the quality of
the model (since all rows containing NAs are dropped from the training data).}

\item{geovars}{Variables containing labels and IDs of the spatial units.}

\item{training_prop}{Optional argument to define a share of observations to be randomly kept in the training
data. It generates a training dataset by excluding the inverse proportion from the training data.}

\item{...}{Optional parameters that can be passed to the \link[caret:train]{caret::train()} function.}
}
\value{
A train object.
}
\description{
This function can be used to train the model for the prediction of one vote
based on a number of past vote results. It uses the machine learning models
available in the caret package.
}
\examples{

train_prediction_model("Eidg1", vote_data, to_exclude_vars = "Kant1")

}
