% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_funs.R
\name{get_differences}
\alias{get_differences}
\title{Calculation of voter turnout differences between all vote topics ("Vorlagen")}
\usage{
get_differences(df, vorl1, vorl2, geo_cols = c("gemwkid", "gemeinde"))
}
\arguments{
\item{df}{A data frame containing the municipality ID and voter turnout for various vote topics,
with each vote topic in a separate column.}

\item{vorl1}{A vector of proposal IDs corresponding to column names in the dataset (e.g., "eidg1", "kant2").}

\item{vorl2}{A vector of proposal IDs (e.g., "eidg1", "kant2") for comparison with `vorl1`.}
}
\value{
A tibble containing voter turnout differences between all possible combinations of vote topics.
}
\description{
Calculation of voter turnout differences between all vote topics ("Vorlagen")
}
\examples{
library(tibble)

testdata <- tibble(gemwkid = c(13,49,41,43,44),
eidg1 = c(60.90,61.18,65.27,55.36,57.68),
eidg2 = c(62.16,62.54,66.95,56.65,58.68),
kant1 = c(57.73,60.27,63.31,51.93,54.49))

single_difference(testdata,"eidg1","eidg2")

 # generate combinations
combinations <-as.data.frame(t(combn(c("eidg1","eidg2","kant1"),2)))

# calculate all possible differences between columns

get_differences(testdata,combinations$V1,combinations$V2)

}
