% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{is_outlier_double_mad}
\alias{is_outlier_double_mad}
\title{Detect Outliers via Median Absolute Deviation from the Median for asymmetric distributions}
\usage{
is_outlier_double_mad(value, thres = 3.5)
}
\arguments{
\item{value}{variable of interest}

\item{thres}{z-score threshold (defaults to 3.5, which is a popular choice).}
}
\value{
logical vector
}
\description{
Outlier detection based on MAD for asymetric distributions. Calculates separate MADs for each half of the distribution.
Median Absolute Deviation is a robust normalization unit based on median as a population center.
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

isnt_outlier_double_mad(x)
}
