% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{is_outlier_turkey}
\alias{is_outlier_turkey}
\title{Detect outliers using turkey's fences}
\usage{
is_outlier_turkey(x, k = 1.5, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{k}{Multiplier for the IQR to set outlier boundaries. Higher values widen the range; default is 1.5.}

\item{na.rm}{if TRUE, removes NA values before calculations. Default is TRUE.}
}
\value{
A logical vector.
}
\description{
Outlier detection based on turkey's fences. Tukey’s fences is a technique used in box plots. The non-outlier range is defined as
\link{Q1−k(Q3−Q1), Q3+k(Q3−Q1)}, where Q1 and Q3 are the lower and upper quartiles respectively and k - some non-negative constant
(popular choice is 1.5).
}
\examples{
x <- c(1, 2, 3, 3, 4, 4, 4, 5, 5.5, 6, 6, 6.5, 7, 7, 7.5, 8, 9, 12, 52, 90)

is_outlier_turkey(x)

}
