% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funs.R
\name{outlier_range}
\alias{outlier_range}
\title{Get boundaries beyond which a value is an outlier via Median Absolute Deviation from the Median for asymmetric distributions}
\usage{
outlier_range(value, thres = 3.5)
}
\arguments{
\item{value}{variable of interest}

\item{thres}{z-score threshold (defaults to 3.5, which is a popular choice).}
}
\value{
tibble with numeric range
}
\description{
Get boundaries beyond which a value is an outlier via Median Absolute Deviation from the Median for asymmetric distributions
}
