% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{predict_single_vote}
\alias{predict_single_vote}
\title{Train a model for a specific vote}
\usage{
predict_single_vote(
  x,
  traindata,
  testdata = NULL,
  method = "bagEarth",
  trControl = NULL,
  to_exclude_vars = NULL,
  geovars = c("gemeinde", "v_gemwkid"),
  testprop = NA,
  ...
)
}
\arguments{
\item{x}{column name of the dependent variable}

\item{traindata}{data used to train the model containing the dependent variable and the predictor-columns}

\item{testdata}{optional dataset structured identically as the trainingdataset on which the prediction should be run. Defaults to NULL, which entails that the prediction is run on the trainingdataset.}

\item{method}{method available in the caret-package which should be used for the prediction}

\item{trControl}{parameters to tune the model}

\item{to_exclude_vars}{variables that should be excluded from the model}

\item{geovars}{variables containing labels and ids of the spatial units}

\item{testprop}{optional argument to generate a training dataset by splitting the dataset (testprop=share of observations to be randomly kept)}

\item{...}{optional parameters that can be passed to the caret::train function}
}
\value{
data.frame
}
\description{
Train a model for a specific vote
}
\examples{

predict_single_vote("Eidg1",votedata, to_exclude_vars = "Kant1")

}
