% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{predict_votes}
\alias{predict_votes}
\title{Run predictions for multiple votes}
\usage{
predict_votes(
  x,
  traindata,
  testdata = NULL,
  method = "svmRadial",
  trControl = NULL,
  exclude_votes = FALSE,
  geovars = c("gemeinde", "v_gemwkid"),
  testprop = NA,
  ...
)
}
\arguments{
\item{x}{Column names of the dependent variables.}

\item{traindata}{Data used to train the model containing the dependent variable and the predictor columns.}

\item{testdata}{Optional dataset structured identically as the training dataset on which the prediction
should be run. Defaults to NULL, which entails that the prediction is run on the training dataset.}

\item{method}{A string specifying which classification or
regression model to use. Possible values are found using
\code{names(getModelInfo())}. See
\url{http://topepo.github.io/caret/train-models-by-tag.html}. A
list of functions can also be passed for a custom model
function. See
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}
for details.}

\item{trControl}{A list of values that define how this function
acts. See \code{\link[caret]{trainControl}} and
\url{http://topepo.github.io/caret/using-your-own-model-in-train.html}.
(NOTE: If given, this argument must be named.)}

\item{exclude_votes}{If set to TRUE, the variables to be predicted will be excluded from each others models.
This can make sense on a vote Sunday due to differences in the counting processes.}

\item{geovars}{Variables containing labels and IDs of the spatial units.}

\item{testprop}{Optional argument to generate a training dataset by splitting the dataset (testprop = share of observations to be randomly kept).}

\item{...}{Optional parameters that can be passed to the \link[caret:train]{caret::train()} function.}
}
\value{
data.frame
}
\description{
This function can be used to predict the outcome of multiple votes based on a number of past vote results.
It uses the machine learning models available in the caret package.
}
\examples{

predict_votes(c("Eidg1","Kant1"), votedata, exclude_votes=TRUE)
}
