% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_funs.R
\name{RMSE}
\alias{RMSE}
\title{Calculate RMSE}
\usage{
RMSE(m, o)
}
\arguments{
\item{m}{predicted value (fitted by modelling)}

\item{o}{oserved 'true' value}
}
\value{
numeric value
}
\description{
Root Mean Square Error (RMSE) = standard deviation of the residuals (prediction errors).
}
\examples{
library(dplyr)
library(tidyr)

pred_data  <- predict_votes(c("Eidg1","Kant1"), votedata, exclude_votes=TRUE)

pred_data \%>\%
drop_na() \%>\%
group_by(vorlage) \%>\%
summarize(rmse=RMSE(pred,real))

}
