% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{votedata}
\alias{votedata}
\title{Test data containing past results and incomplete results for two votes}
\format{
A dataframe with 171 rows and 79 columns.
\describe{
  \item{gemeinde}{municipality name}
  \item{v_gemwkid}{vote id}
  \item{Eidg1}{Yes-Share of an on going federal vote}
  \item{Kant1}{Yes-Share of an on going  cantonal vote}
  \item{v_....}{yes shares of past votes}
}
}
\usage{
votedata
}
\description{
Test data containing past results and incomplete results for two votes
}
\keyword{datasets}
